Exclude game info from GTK3 board preview
In the preferences dialog, the board preview is shown with game info
below it that is unnecessary. This area is not aligned properly and
overlaps with the graphical board display.

Adjusting this required a change in draw_game_info(): I've separated
board data init code from the actual gtk rendering.

Now we can draw the board without its info below it, to simplify
alternative board rendering such as in the preferences.