Add option to to re-evaluate lower ply moves
If fAnalyse is true we don't want to have a move evaluated at a
lower ply to appear better than a one evaluated at a higher ply:
re-evaluate it at the ply of the previous move

Apply this in analysis and hint, that is when we don't need only
the best move but are interested in errors cost.

Apply it after regular move filter but before reevaluating the actual
move when the latter is out of the move filter.